+++ packages/appkit-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ton/appkit-react
 
+## 0.0.6-alpha.1
+
+### Patch Changes
+
+- @ton/appkit@0.0.5-alpha.1
+
+## 0.0.6-alpha.0
+
+### Patch Changes
+
+- @ton/appkit@0.0.5-alpha.0
+
 ## 0.0.5
 
 ### Patch Changes